commit modified files

+++ src/main/java/com/example/skillcatchupsep/api/repository/PrefixRepository.java
@@ -1,5 +1,6 @@
 package com.example.skillcatchupsep.api.repository;
 
+import com.example.skillcatchupsep.api.PrefixEntity;
 import org.seasar.doma.Dao;
 import org.seasar.doma.Insert;
 import org.seasar.doma.Select;
@@ -13,11 +14,11 @@ import java.util.List;
 public interface PrefixRepository {
 
     @Select
-    public PrefixEntity findById(Long prefixId);
-
+    public PrefixEntity findById(Integer prefixId);
+    
     @Select
     public List<PrefixEntity> findAll();
-
+    
     @Select
     public List<PrefixEntity> findTop30();
